refactored all loops and template files
The loops need to be seperate and preferably under theme-partials/post-*
files for future cleaning. The format in other template files is not as
important for cleanup but for clarity a standardized/readable format has
been implemented in all files.

Fixed other theme errors. Theme should now be in a usable state for
integration.